+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Features
 - **go:** test function now never cache
+- **golang:** added debug current function
 - **golang:** change mappings for nvim goc
 - **nvim-tree:** update config
 - **nvim-tree:** ignore notify window